[Concat] Refactor kernel functions (#525)
* Header files

* [Concat] Refactor kernel fns

* Format

* Fix cast and add init kernel

* Fix Internal

* Remove static

* Remove hash utils include
