+++ include/flexflow/ops/concat.h
@@ -1,23 +1,12 @@
 #ifndef _FLEXFLOW_CONCAT_H
 #define _FLEXFLOW_CONCAT_H
 
-#include "flexflow/accessor.h"
-#include "flexflow/device.h"
-#include "flexflow/fftype.h"
 #include "flexflow/layer.h"
 #include "flexflow/node.h"
-#include "flexflow/op_meta.h"
 #include "flexflow/operator.h"
 #include "flexflow/ops/concat_params.h"
 
 namespace FlexFlow {
-
-class ConcatMeta : public OpMeta {
-public:
-  ConcatMeta(FFHandler handle) : OpMeta(handle){};
-  int legion_axis;
-  char op_name[MAX_OPNAME];
-};
 
 class Concat : public Op {
 public:
@@ -48,7 +37,6 @@ public:
                            std::vector<Legion::PhysicalRegion> const &regions,
                            Legion::Context ctx,
                            Legion::Runtime *runtime);
-  void init_meta(ConcatMeta *meta) const;
   static void forward_task(Legion::Task const *task,
                            std::vector<Legion::PhysicalRegion> const &regions,
                            Legion::Context ctx,
@@ -57,26 +45,6 @@ public:
                             std::vector<Legion::PhysicalRegion> const &regions,
                             Legion::Context ctx,
                             Legion::Runtime *runtime);
-  static void forward_kernel(GenericTensorAccessorW const &output,
-                             GenericTensorAccessorR const *inputs,
-                             int num_inputs,
-                             int axis,
-                             ffStream_t stream);
-  static void forward_kernel_wrapper(ConcatMeta const *m,
-                                     GenericTensorAccessorW const &output,
-                                     GenericTensorAccessorR const *inputs,
-                                     int num_inputs,
-                                     int axis);
-  static void backward_kernel(GenericTensorAccessorR const &output_grad,
-                              GenericTensorAccessorW const *input_grads,
-                              int num_inputs,
-                              int axis,
-                              ffStream_t stream);
-  static void backward_kernel_wrapper(ConcatMeta const *m,
-                                      GenericTensorAccessorR const &output_grad,
-                                      GenericTensorAccessorW const *input_grads,
-                                      int num_inputs,
-                                      int axis);
   bool measure_operator_cost(Simulator *sim,
                              MachineView const &pc,
                              CostMetrics &cost_metrics) const override;
